fix: moving the challenger blocks to experiments folder
